requirements.txt
asgiref==3.5.2
Django==4.0.4
django-admin-rangefilter==0.8.4
psycopg2==2.9.3
sqlparse==0.4.2
